LayoutTests/platform/ios-simulator/fast/events/dropzone-004-expected.txt

CONSOLE MESSAGE: line 58: TypeError: undefined is not a function (evaluating 'eventSender.beginDragWithFiles(["resources/file-for-prevent-drag-to-navigate.html"])')
This test checks that drag-and-drop support works with files.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Drop the file onto me.

Expects dropEffect 
